+++ CHANGES.rst
@@ -1,3 +1,9 @@
+..
+    Copyright 2022 Rafe Kaplan
+    SPDX-License-Identifier: Apache-2.0
+
+    Updated: 2022-06-06
+
 v0.4.0
 ======
 - Provide alternate location for configuration file.